requirements.txt
sentence_transformers
weaviate-client